requirements.txt
streamlit>=1.37
python-dotenv>=1.0
openai>=1.43
numpy>=1.24
scikit-learn>=1.3
rank-bm25>=0.2.2
sentence-transformers>=3.0
transformers>=4.43
torch>=2.2; sys_platform != 'darwin' or platform_machine != 'arm64'
torch>=2.1; sys_platform == 'darwin' and platform_machine == 'arm64'
beautifulsoup4>=4.12
requests>=2.32
python-docx>=1.1